[BUGFIX] just react on some clearCache commands

+++ Classes/Hooks/T3libTcemainHook.php
@@ -2,9 +2,11 @@
 
 class tx_dyncss_Hooks_T3libTcemainHook {
 	function clearCachePostProc($params, &$pObj) {
-		t3lib_div::rmdir(
-			PATH_site . 'typo3temp/Cache/Data/DynCss',
-			TRUE
-		);
+		if(isset($params['cacheCmd']) && $params['cacheCmd'] = 'pages') {
+			t3lib_div::rmdir(
+				PATH_site . 'typo3temp/Cache/Data/DynCss',
+				TRUE
+			);
+		}
 	}
 }
